Manage errors in > < >> << cases and pipes duplication
